pyverbs: Fix randrange deprecation warning
[ Upstream commit ff5b2ac83cfb948309cbd0b35931e666c7d514d4 ]

Second argument to randrange() call is not integer as such every time
rdma-core tests were run, the following deprecation warnings were
printed.

./tests/test_device.py:284: DeprecationWarning: non-integer arguments to randrange()
have been deprecated since Python 3.10 and will be removed in a subsequent version
  dm_len = random.randrange(u.MIN_DM_SIZE, self.attr_ex.max_dm_size/2,

Let's cast that argument to integer.

Fixes: 259fab3903f0 ("tests: Avoid large allocation attempts of device memory")

+++ tests/test_device.py
@@ -253,7 +253,7 @@ class DMTest(PyverbsAPITestCase):
         """
         test ibv_alloc_dm()
         """
-        dm_len = random.randrange(u.MIN_DM_SIZE, self.attr_ex.max_dm_size/2,
+        dm_len = random.randrange(u.MIN_DM_SIZE, int(self.attr_ex.max_dm_size/2),
                                   u.DM_ALIGNMENT)
         dm_attrs = u.get_dm_attrs(dm_len)
         with d.DM(self.ctx, dm_attrs):
@@ -263,7 +263,7 @@ class DMTest(PyverbsAPITestCase):
         """
         test ibv_free_dm()
         """
-        dm_len = random.randrange(u.MIN_DM_SIZE, self.attr_ex.max_dm_size/2,
+        dm_len = random.randrange(u.MIN_DM_SIZE, int(self.attr_ex.max_dm_size/2),
                                   u.DM_ALIGNMENT)
         dm_attrs = u.get_dm_attrs(dm_len)
         dm = d.DM(self.ctx, dm_attrs)
@@ -299,7 +299,7 @@ class DMTest(PyverbsAPITestCase):
         """
         Test calling ibv_free_dm() twice
         """
-        dm_len = random.randrange(u.MIN_DM_SIZE, self.attr_ex.max_dm_size/2,
+        dm_len = random.randrange(u.MIN_DM_SIZE, int(self.attr_ex.max_dm_size/2),
                                   u.DM_ALIGNMENT)
         dm_attrs = u.get_dm_attrs(dm_len)
         dm = d.DM(self.ctx, dm_attrs)
@@ -310,7 +310,7 @@ class DMTest(PyverbsAPITestCase):
         """
         Test writing to the device memory
         """
-        dm_len = random.randrange(u.MIN_DM_SIZE, self.attr_ex.max_dm_size/2,
+        dm_len = random.randrange(u.MIN_DM_SIZE, int(self.attr_ex.max_dm_size/2),
                                   u.DM_ALIGNMENT)
         dm_attrs = u.get_dm_attrs(dm_len)
         with d.DM(self.ctx, dm_attrs) as dm:
@@ -324,7 +324,7 @@ class DMTest(PyverbsAPITestCase):
         """
         Test writing to the device memory with bad offset and length
         """
-        dm_len = random.randrange(u.MIN_DM_SIZE, self.attr_ex.max_dm_size/2,
+        dm_len = random.randrange(u.MIN_DM_SIZE, int(self.attr_ex.max_dm_size/2),
                                   u.DM_ALIGNMENT)
         dm_attrs = u.get_dm_attrs(dm_len)
         with d.DM(self.ctx, dm_attrs) as dm:
@@ -345,7 +345,7 @@ class DMTest(PyverbsAPITestCase):
         """
         Test reading from the device memory
         """
-        dm_len = random.randrange(u.MIN_DM_SIZE, self.attr_ex.max_dm_size/2,
+        dm_len = random.randrange(u.MIN_DM_SIZE, int(self.attr_ex.max_dm_size/2),
                                   u.DM_ALIGNMENT)
         dm_attrs = u.get_dm_attrs(dm_len)
         with d.DM(self.ctx, dm_attrs) as dm:
